Fixed execution error on windows
The ingress cant be accessed on windows only on Linux (and probably MacOS). I tried a lot of different approaches and looked into this toppic online. It seems that minikube on Windows cant properly expose the ingress to the outside world on windows. On Linux on the other hand it is exposed by default. The workaround is documented in k8s/README.md. It is to use the loadbalancer service for the frontend instead of the ingress to access the application.