spark-submit now working
The spark-submit script is used by an external pod to start spark jobs. The spark master node can be reached and the job is deployed on the spark cluster. Also the web GUI shows the running jobs and the connected spark workers.
Next step: Use spark specific functions to deploy the ML code on the spark cluster. Also communicatation with kafka needs to be implemented.