Cassandra connection with spark
The configurations to connect cassandra to spark require parameters in the Spark/Dockerfile. They are parsed to the spark-submit script to prepare the spark cluster for the connection. It also extends the functionality of spark dataframes with some cassandra specific functions.

The setup requires the user to run some commands before using skaffold dev.